renamed pages to their new correct names
