+++ packages/react-frontend/src/pages/artists.jsx
